requirements.txt
discord.py==2.3.2
Flask
aiohttp
psutil
python-dotenv
pymongo[srv]==4.6.1
motor==3.3.2